feat(maze.py): remove entrance and exit
After Cell instance have been initialized we call the new method
`_break_entrance_and_exit()`  that is responsible for removing the needed
walls (setting it to False) for the top left and bottom right cells
respectively. We also added an Enum based class that should eventually
be used to set and find out wall information.