feat(maze.py): replace maze generation strategy
With this change, the previous approach that would attach the maze
generation with the cell rendering code was discarded in favor of a
graph, DFS algorithm, to ensure no cycles and connectivity on the
generated maze. This new approach should be more testable and faster to
work with, as for solving we can use the same data member.